+++ src/Epinova.ElasticSearch.Core/Models/JsonNames.cs
@@ -16,6 +16,7 @@
         public const string DidYouMean = "dym";
         public const string Doc = "doc";
         public const string DocCount = "doc_count";
+        public const string Dynamic = "dynamic";
         public const string Error = "error";
         public const string Field = "field";
         public const string FieldData = "fielddata";